Cast ternary mode arg to u8 in Sky_func03 (99.674->99.905)
skyFn_80088c94's mode param is u8-narrowed; casting the ternary
result reproduces the li r0,K; clrlwi r4,r0,24 staging.

+++ src/main/sky.c
@@ -3179,7 +3179,7 @@ void Sky_func03(int a, int b, u8* cfg)
         }
         if (((Sky2Config*)cfg)->unk5D != 0)
         {
-            skyFn_80088c94(mask, ((Sky2Config*)cfg)->unk5D > 2 ? 1 : 0);
+            skyFn_80088c94(mask, (u8)(((Sky2Config*)cfg)->unk5D > 2 ? 1 : 0));
         }
         vis = ((Sky2Config*)cfg)->unk56;
         for (i = 0; i < 2; i++)
